refactor(server): extract shared health status/component types
